refactor: 불필요 Direction 메서드 제거 및 테스트 수정
