+++ src/main/java/chess/domain/piece/Knight.java
@@ -2,6 +2,7 @@ package chess.domain.piece;
 
 import java.util.List;
 import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class Knight extends Piece {
 
@@ -27,8 +28,8 @@ public class Knight extends Piece {
     }
 
     private boolean isInRange(Position targetPosition) {
-        List<Position> inRangePosition = Direction.getKnightDirections()
-                .stream()
+        List<Position> inRangePosition = Stream.of(Direction.NNE, Direction.NNW, Direction.SSE, Direction.SSW,
+                        Direction.EEN, Direction.EES, Direction.WWN, Direction.WWS)
                 .filter(direction -> Position.isValidPosition(Position.createNextPosition(position, direction)))
                 .map(direction -> Position.createNextPosition(position, direction))
                 .collect(Collectors.toList());
